Update map.js
I updated the file significantly. I replaced some variable names with more descriptive ones to make the code easier to comprehend. The map-center coordinates are now being retrieved from the database of the specified city, and the correct city is displayed when selected.
I also revised the code for adding the various layers to make it more efficient and comprehensible. I created an object for styling the different layers and used a for loop to apply the appropriate styling to each geojson "zone," as opposed to the previous approach.
Changed the map